+++ metadata.rb
@@ -1,8 +1,8 @@
-name             'dev-build-essential'
+name             'build-essential-dev'
 maintainer       'three18ti'
 maintainer_email 'three18ti@gmail.com'
 license          'All rights reserved'
-description      'Installs/Configures dev-build-essential'
+description      'Installs/Configures build-essential-dev'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'